Simplify src/runtime/engines tree
