--- src/runtime/engines/all/all_engines.go
@@ -1,10 +1,0 @@
-// Copyright (c) 2018, Sylabs Inc. All rights reserved.
-// This software is licensed under a 3-clause BSD license. Please consult the
-// LICENSE file distributed with the sources of this project regarding your
-// rights to use or distribute this software.
-
-package all
-
-import (
-	_ "github.com/singularityware/singularity/src/runtime/engines/singularity" // singularity engine
-)
