+++ src/runtime/startup/scontainer.go
@@ -9,8 +9,7 @@ package main
 #include <sys/types.h>
 #include "startup/c/wrapper.h"
 */
-// #cgo CFLAGS: -I../.. -I../../c/lib
-// #cgo LDFLAGS: -L../../../../builddir/lib/ -lruntime -luuid
+// #cgo CFLAGS: -I..
 import "C"
 
 import (
@@ -23,7 +22,6 @@ import (
 	"github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/singularityware/singularity/src/pkg/sylog"
 	"github.com/singularityware/singularity/src/runtime/engines"
-	_ "github.com/singularityware/singularity/src/runtime/engines/all"
 )
 
 func bool2int(b bool) uint8 {
